requirements.txt
streamlit
pillow
pytesseract
openai